Fixed conversion warning

+++ Source/Systems.cpp
@@ -61,7 +61,7 @@ namespace ChildrensTears {
         auto render    = &coord.getComponent<RenderComponent>(entity);
 
         // Update position of the sprite
-        render->sprite.setTextureRect({0, 0, transform->size.x, transform->size.y});
+        render->sprite.setTextureRect({0, 0, static_cast<int>(transform->size.x), static_cast<int>(transform->size.y)});
         render->sprite.setPosition(transform->position.x, transform->position.y);
 
         // Draw to render area
